+++ CMakeLists.txt
@@ -528,6 +528,7 @@ set(scylla_sources
     partition_version.cc
     querier.cc
     query.cc
+    query_ranges_to_vnodes.cc
     query-result-set.cc
     raft/fsm.cc
     raft/log.cc
