+++ configure.py
@@ -824,6 +824,7 @@ scylla_core = (['replica/database.cc',
                 'service/priority_manager.cc',
                 'service/migration_manager.cc',
                 'service/storage_proxy.cc',
+                'query_ranges_to_vnodes.cc',
                 'service/paxos/proposal.cc',
                 'service/paxos/prepare_response.cc',
                 'service/paxos/paxos_state.cc',
