+++ test/boost/storage_proxy_test.cc
@@ -15,6 +15,7 @@
 #include "test/lib/mutation_source_test.hh"
 #include "test/lib/result_set_assertions.hh"
 #include "service/storage_proxy.hh"
+#include "query_ranges_to_vnodes.hh"
 #include "partition_slice_builder.hh"
 #include "schema_builder.hh"
 
@@ -42,7 +43,7 @@ SEASTAR_TEST_CASE(test_get_restricted_ranges) {
 
             auto check = [&s](locator::token_metadata_ptr tmptr, dht::partition_range input,
                               dht::partition_range_vector expected) {
-                service::query_ranges_to_vnodes_generator ranges_to_vnodes(tmptr, s, {input});
+                query_ranges_to_vnodes_generator ranges_to_vnodes(tmptr, s, {input});
                 auto actual = ranges_to_vnodes(expected.size());
                 if (!std::equal(actual.begin(), actual.end(), expected.begin(), [&s](auto&& r1, auto&& r2) {
                     return r1.equal(r2, dht::ring_position_comparator(*s));
